pyproject.toml
[tool.poetry]
name = "djangoauthproject"
version = "0.1.0"
description = ""
authors = ["MuhammedYO1307 <mahmud150@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.5"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.10"
redis = "^6.4.0"
pyjwt = "^2.10.1"
dj-database-url = "^3.0.1"
python-dotenv = "^1.1.1"
djangorestframework-simplejwt = "^5.5.1"
drf-yasg = "^1.21.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
